+++ compiler/qsc_partial_eval/src/management.rs
